+++ src/audio/tdfb/tdfb_hifiep.c
@@ -5,11 +5,10 @@
 // Author: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>
 
 #include <sof/common.h>
-#include <user/tdfb.h>
 #include <sof/audio/audio_stream.h>
-#include <sof/audio/tdfb/tdfb_comp.h>
 #include <user/fir.h>
-#include <user/tdfb.h>
+
+#include "tdfb_comp.h"
 
 #if TDFB_HIFIEP
